prowler/pyproject.toml
[project]
name = "prowler"
version = "0.1.0"
description = ""
authors = [
    {name = "unfortunatelygeek",email = "aditiiirao15@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse (>=1.4.0,<2.0.0)",
    "structlog (>=25.1.0,<26.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "google-generativeai (>=0.8.4,<0.9.0)",
    "jedi (>=0.19.2,<0.20.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
grpcio = "1.60.1"

